Move Factory Test to Demos
The factory_test.hpp file is designed solely for the SJTwo board. Given
that it is in the library folder it would seem that it would be generic
in some fashion. Since it is not, Its code has been moved to the factory
test demo project and its original header deleted.

Resolves #735
